+++ src/engine/lmr.rs
@@ -2,7 +2,7 @@ const BASE_REDUCTION: f64 = 0.75;
 const DIVISOR: f64 = 2.25;
 
 pub struct LMRTable {
-    pub table: [[i16; 64]; 64],
+    pub table: [[i32; 64]; 64],
 }
 
 impl LMRTable {
@@ -15,7 +15,7 @@ impl LMRTable {
             for m in 0..64 {
                 let ld = f64::ln(d as f64);
                 let lp = f64::ln(m as f64);
-                lmr.table[d][m] = (BASE_REDUCTION + ld * lp / DIVISOR) as i16;
+                lmr.table[d][m] = (BASE_REDUCTION + ld * lp / DIVISOR) as i32;
             }
         }
 
@@ -23,7 +23,7 @@ impl LMRTable {
     }
 
     #[must_use]
-    pub fn reduction(&self, depth: i16, move_count: i16) -> i16 {
+    pub fn reduction(&self, depth: i32, move_count: i32) -> i32 {
         let d = (depth.min(63)) as usize;
         let c = (move_count.min(63)) as usize;
