+++ src/engine/tt.rs
@@ -75,7 +75,7 @@ impl TT {
         score: i16,
         depth: u8,
         flag: TTFlag,
-        ply: u8,
+        ply: i32,
     ) {
         let target_index = self.index(key);
         let target = self.entries[target_index];
@@ -95,10 +95,10 @@ impl TT {
     }
 
     #[must_use]
-    pub fn score_to_tt(&self, score: i16, ply: u8) -> i16 {
-        if score >= TB_WIN_IN_PLY {
+    pub fn score_to_tt(&self, score: i16, ply: i32) -> i16 {
+        if score >= TB_WIN_IN_PLY as i16 {
             score + ply as i16
-        } else if score <= TB_LOSS_IN_PLY {
+        } else if score <= TB_LOSS_IN_PLY as i16 {
             score - ply as i16
         } else {
             score
@@ -106,10 +106,10 @@ impl TT {
     }
 
     #[must_use]
-    pub fn score_from_tt(&self, score: i16, ply: u8) -> i16 {
-        if score >= TB_WIN_IN_PLY {
+    pub fn score_from_tt(&self, score: i16, ply: i32) -> i16 {
+        if score >= TB_WIN_IN_PLY as i16 {
             score - ply as i16
-        } else if score <= TB_LOSS_IN_PLY {
+        } else if score <= TB_LOSS_IN_PLY as i16 {
             score + ply as i16
         } else {
             score
